+++ src/Symfony/Component/Messenger/Bridge/MongoDB/CHANGELOG.md
@@ -4,4 +4,4 @@ CHANGELOG
 6.3
 ---
 
- * Introduced the Redis bridge.
+ * Introduced the MongoDB bridge.